requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.0.3
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
email-validator==2.1.1
